+++ sandpack-react/src/contexts/sandpackContext.tsx
@@ -352,6 +352,13 @@ export class SandpackProviderClass extends React.PureComponent<
         showErrorScreen: !this.errorScreenRegistered.current,
         showLoadingScreen: !this.loadingScreenRegistered.current,
         reactDevTools: this.state.reactDevTools,
+        customNpmRegistries: this.props.customSetup?.npmRegistries?.map(
+          (config) =>
+            ({
+              ...config,
+              proxyEnabled: false, // force
+            } ?? [])
+        ),
       }
     );