+++ sandpack-react/src/themes/Theme.stories.tsx
@@ -9,6 +9,14 @@ const stories = storiesOf("presets/Themes", module);
 
 Object.keys(SANDPACK_THEMES).forEach((themeName) =>
   stories.add(themeName, () => (
-    <Sandpack theme={themeName as keyof typeof SANDPACK_THEMES} />
+    <Sandpack
+      options={{
+        showLineNumbers: true,
+        showInlineErrors: true,
+        showNavigator: true,
+        showTabs: true,
+      }}
+      theme={themeName as keyof typeof SANDPACK_THEMES}
+    />
   ))
 );