+++ sandpack-themes/src/index.ts
@@ -3,3 +3,4 @@ export { githubLight } from "./githubLight";
 export { monokaiPro } from "./monokaiPro";
 export { nightOwl } from "./nightOwl";
 export { sandpackDark } from "./sandpackDark";
+export { dracula } from "./dracula";